requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
html5lib==1.1
idna==2.10
psycopg2==2.9.9
python-dateutil==2.8.1
requests==2.31.0
six==1.16.0
soupsieve==2.2.1
urllib3==2.1.0
webencodings==0.5.1
